[refactor/DATEPICK-227]: BaseEntity 삭제 및 id 값들 모두 entity로 옮기기 + Entity 네이밍, 타입 수정

+++ src/main/java/app/hdj/datepick/domain/course/dto/CourseRequest.java
@@ -23,6 +23,6 @@ public class CourseRequest {
     @NotNull
     private Boolean isPrivate;
 
-    private List<@Positive Byte> tagIds;
+    private List<@Positive Integer> tagIds;
 
 }
